[Go] make the two Gemini generators more similar
Reorganize the code in the google AI and Vertex plugins to be
more similar.

No behavior changes.
